Create data package and move api inside
