+++ android/CodeTest/app/src/main/java/com/codetest/main/data/api/LocationApiService.kt
@@ -1,4 +1,4 @@
-package com.codetest.main.api
+package com.codetest.main.data.api
 
 import com.codetest.main.KeyUtil
 import com.google.gson.JsonObject
