+++ android/CodeTest/app/src/main/java/com/codetest/main/di/AppModule.kt
@@ -2,7 +2,7 @@ package com.codetest.main.di
 
 import android.content.Context
 import com.codetest.main.KeyUtil
-import com.codetest.main.api.LocationApiService
+import com.codetest.main.data.api.LocationApiService
 import com.google.gson.Gson
 import com.google.gson.GsonBuilder
 import dagger.Module
